Move util file inside migration dir
