+++ scripts/migration-project_15_0_0/perf/get_submissions_info.py
@@ -5,7 +5,7 @@ import logging
 
 import requests
 
-from util import load_json, write_json
+from migration.util import load_json, write_json
 
 FILENAME = 'submission-envelopes.json'
